Code quality fixes
Parser.parse method has many if statements that throws a BaymaxException
with a unique message.

Refactored the method to improve readability.

Extracting the if and throw behaviour into a separate method, and the
statement in each switch case is more readable.

Let's,
* create a method to check if there are arguments and throw an exception
if there are none with a unique message for each command.
* create a method to get the index of the desired string that indicates
the start of a parameter for the command and throw an exception if it
cannot be found with a unique message for each command.